Bandwidth-aware profile selection + design-wide defect-class audit fixes (systemic currency mixing, rate-card/adapter correctness, resiliency sentinels)
The owner asked whether fixes live in the underlying engines (shared by
all consumers) or just the MCP layer, and for THOROUGH fixes. Confirmed
the MCP layer is a thin shell over the engines, then ran a design-wide
audit of the nine defect classes fixed in the optimizer/wizard against
EVERY design component (value-engineering, rate cards, provider
adapters, peering/resiliency, geo, registry). 34 findings; the real ones
are fixed here. All fixes are in the shared engine/model code, not MCP.

BANDWIDTH-AWARE SERVICE-PROFILE SELECTION (provider-general): the wizard
picked one profile per (provider, metro) with no regard to bandwidth, so
an AWS hosted profile (tiers <=500) was pinned to a 3000 Mbps connection
(live EQ error). The optimizer now carries every matching profile's
capability (ServiceProfileOption: tiers, allowCustom, per-metro ceiling)
forward; the wizard selects, once the connection bandwidth is known, a
profile whose tiers cover it (dedicated when > hosted max), or emits a
precise actionable error and no unbuildable connection. covers() mirrors
PlanValidator.checkProfile so selector and validator agree.

SYSTEMIC CURRENCY MIXING (audit HIGH C9) — the headline financial bug,
present at FIVE cost sites: TcoEngine, SavingsCalculatorEngine,
MetroOptimizerEngine.estimateCosts, DeploymentWizardEngine.
estimatePricing, PlanValueAssessment all added USD egress to the live
Fabric MRC in the account/metro currency (EUR/GBP/SGD for EMEA/APAC) and
labelled the sum one currency — every non-USD estimate silently wrong.
New CurrencyReconciler is the single rule: never combine quotes of
differing currency; reconcile only when all share one, else surface
per-currency subtotals + a reason and mark the total unpriced/mixed (no
fabricated FX). Wired at all five sites.

RATE-CARD CORRECTNESS: EquinixRateCard metro match was a 2-letter
substring (ME/LA/DE false-positive inside price words) -> whole-token;
a phantom $0 authoritative quote from unparsed charges -> Optional.empty;
a transient fetch failure cached forever -> memoize only success;
ReferenceRateCard above-top-tier bandwidth silently priced at the top
flat rate -> unpriced/tagged.

PROVIDER ADAPTERS: Azure never followed NextPageLink and OCI single-GET
(first-page-only pricing) -> paginate; GCP over-broad egress-SKU pick ->
deterministic; per-GiB vs per-GB corrected.

PEERING / RESILIENCY: unknown distance was 0 km -> a fabricated CRITICAL
"same-site" rating + poisoned resiliency score; now modelled absent
(NaN) and stated honestly. Silent Fabric-geo / facility / opportunity
swallows -> surfaced per the skip-and-call-out rule. Dead levers
(includeFabricConnections/includeCapacity/includePolicies) resolved;
integer-Gbps truncation, int-overflow capacity sum, and UNKNOWN-region-
as-equal fixed.

HONESTY CLEANUPS: latency endpoint renders the real metro code (not the
UNKNOWN enum); MetroOptimizer.monthlyBudget javadoc corrected (reported-
against, not enforced); Fabric.authenticate registry-warm failure logged
instead of silently swallowed.

Fair note: these components were well-built (guards on reference figures,
honest disclaimers); the bugs were the same class each time — guards on
the cosmetic paths but not the load-bearing live ones, and absent-value
sentinels (0 km, $0, floor tier) leaking into agent-facing numbers. Only
bite against live/international data, which is why the demo missed them.

Gate green: 1013 default + 1207 wiremock + coverage.

+++ src/main/java/api/equinix/javasdk/design/peering/model/FailoverPath.java
@@ -50,7 +50,7 @@ public class FailoverPath {
 
     ConnectivityType connectivityType;
 
-    int ixCapacityMbps;
+    long ixCapacityMbps;
 
     boolean routeServerAvailable;
